backend/requirements.txt
boto3==1.28.64
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-mock==3.12.0
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1 